+++ lib/ssd1306/ssd1306.h
@@ -15,18 +15,18 @@ public:
 
     void begin() final;
     void clear() final;
-    void clear_pixel(uint8_t x, uint8_t y) final;
-    void clear_area(uint8_t x_start, uint8_t y_start, uint8_t x_end, uint8_t y_end) final;
+    void clear_pixel(int16_t x, int16_t y) final;
+    void clear_area(int16_t x_start, int16_t y_start, int16_t x_end, int16_t y_end) final;
     void flush() final;
     bool next_segment() final;
-    void draw_pixel(uint8_t x, uint8_t y) final;
-    void draw_bitmap(uint8_t x, uint8_t y, Bitmap&& bitmap) final;
-    void invert_area(uint8_t x_start, uint8_t y_start, uint8_t x_end, uint8_t y_end) final;
-    void draw_h_line(uint8_t y, uint8_t x_start, uint8_t x_end, char linestyle) final;
-    void draw_v_line(uint8_t x, uint8_t y_start, uint8_t y_end, char linestyle) final;
-    void draw_line(uint8_t x_start, uint8_t y_start, uint8_t x_end, uint8_t y_end, char linestyle) final;
-    void draw_circle(uint8_t x, uint8_t y, uint8_t radius) final;
-    void draw_arc(uint8_t x, uint8_t y, uint8_t radius, uint16_t angle_start, uint16_t angle_end) final;
+    void draw_pixel(int16_t x, int16_t y) final;
+    void draw_bitmap(int16_t x, int16_t y, Bitmap&& bitmap) final;
+    void invert_area(int16_t x_start, int16_t y_start, int16_t x_end, int16_t y_end) final;
+    void draw_h_line(int16_t y, int16_t x_start, int16_t x_end, char linestyle) final;
+    void draw_v_line(int16_t x, int16_t y_start, int16_t y_end, char linestyle) final;
+    void draw_line(int16_t x_start, int16_t y_start, int16_t x_end, int16_t y_end, char linestyle) final;
+    void draw_circle(int16_t x, int16_t y, uint8_t radius) final;
+    void draw_arc(int16_t x, int16_t y, uint8_t radius, uint16_t angle_start, uint16_t angle_end) final;
 
 private:
     void command(uint8_t cmd);
